VGA Test image works. Yay
Well, only Red has the full color depth. Blue and Green are basically
black unless they are set to 2'b11. I'll have to figure out why that is
at some point.

The connections on my DAC are not the issue (at least not in a trivial
way). When I probe the top of the color screw terminals with my
oscilloscope, the range looks fine (from 0V to 0.7V with even steps) and
it looks the same for all colors.

The VGA cable does not seem to be the problem since it works fine if I
connect my server to my TV.

Maybe the issue is with the monitor that I used (the right one on my
desk). I should try the left one and maybe my TV to see if the issue
persists.

Maybe the issue is with the solder connection. I should test that by
measuring the resistance from the top of the screw terminal to the end
of the VGA cable

Maybe the issue is that I connected the cables to the connector wrong
(e.g. mixed up blue and blue ground). I should check that before
anything else.