Add an AXI-Lite interface to the sram
This was surprisingly easy. Holy fuck :P

And it'll be super usefuly for the VGA thing as well because it already
does all the arbitration I need (read > write) so I can just give the
VGA core the read end and the uart the write end and they'll hapily work
together like I want them to.