Sram-Axi: Kill performance for the greater good
The greater good is not dropping responses if the other side can't
accept them.

Unfortunately, even with the luxury of being able to accept a request
one cycle after it is issued without sacrificing performance, I'll still
have to implement some sort of buffering when the resposne channel is
stalled. The reason for this is that the sram has an internal latency of
3-4 transactions. This means that I need a four element deep skid fifo
to buffer the read / write responses that were produced by the sram
before we know that the other axi side can't accept them.