Update sram_axi to use skidbuffer
And add some formal properties. The problem with those is that right
now, verifying the sram_axi module in isolation would require me to
assume properties of the sram_* signals which I don't want to do because
I fear making a mistake that would invalidate my results. However, I
cannot verify the thing in conjunction with sram_top because yosys can't
handle the tri state logic of the simulation instances of SB_IO.

I think what I should do is create some assumptions about the sram,
assume them in sram_axi and assert them in sram_ctrl and then try to
formally verify sram_ctrl first and then use that confidence in my
assumptions to verify sram_axi. For this, I should ideally write a
properties file that can be included in both with an ifdef switch that
can switch between assumptions and assertions (I think the giselquist
guy at some point hat a simmilar problem because I noticed this pattern
in his files too)

There is also an sby file for the verification and an sram_axi_top that
tries to integrate the sram block into one module from when I tried to
do verify everything together.