VGA: fix color output timing
Previously, it was lagging behind a bit. This got fixed now but it
relies on the pixel clock being slower than the main clock which isn't
nice.

The color outputs are also registered at the top level now which would
probably a good idea to do for the sync outputs as well to have as
little skew between them as possible